更新代码-调整目录、添加数据中心相关代码

+++ src/components/Accordion/index.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useRef, useState } from 'react';
 import './index.less';
-import { AppstoreAddOutlined, DownOutlined, RightOutlined } from '@ant-design/icons';
-import { Menu, Tree } from 'antd';
+import { AppstoreAddOutlined, CheckOutlined, CloseOutlined, DownOutlined, RightOutlined } from '@ant-design/icons';
+import { Input, Menu, Modal, Tree, message } from 'antd';
 import { getAssetsTree} from '@/services/assets/asset';
-import { addAssetTree,deleteAssetTree } from '@/services/assets/asset-tree';
+import { addAssetTree,updateAssetTree,deleteAssetTree } from '@/services/assets/asset-tree';
 import queryString from 'query-string';
 
 
@@ -15,20 +15,25 @@ interface Group {
 interface Props {
   handleClick: (values: any,status:any,index) => Promise<void>;
   assetStatus: Group[];
+  isAutoInitialized:boolean;
 }
 
 export default function Accordions(props: Props) {
   const { assetStatus } = props;
-  const [bindIndex, setBindIndex] = React.useState(0);
-  const [openStatus, setOpenStatus] = useState<number>(0)
-  const [treeData, setTreeData] = useState<[]>();
+  
+  const [treeData, setTreeData] = useState<any>();
+  const [isAutoInitialized, setIsAutoInitialized] = useState<boolean>(props.isAutoInitialized)
   const [showMenu, setShowMenu] = useState(false);
   const status = queryString.parse(location.search).status === undefined ? 0 : parseInt("" + queryString.parse(location.search).status);
   const index = queryString.parse(location.search).index === undefined ? 0 : parseInt("" + queryString.parse(location.search).index);
   const [pageX, setPageX] = useState(0);
+  const [openStatus, setOpenStatus] = useState<number>(status)
+  const [bindIndex, setBindIndex] = React.useState(index);
+  
   const [pageY, setPageY] = useState(0);
-  const [selectNode,setSelectNode]= useState({});
+  const [selectNode,setSelectNode]= useState<any>({});
   const dropdownElement: React.RefObject<HTMLDivElement> = useRef(null);
+  const [curValue, setCurValue] = useState<string>('');
   const menu = (
     <Menu className='sub_menu'
       onClick={event => operateCategory(event)}
@@ -52,12 +57,90 @@ export default function Accordions(props: Props) {
       ]}
     />
   );
+  const addEditProperty = arr => arr.map(item => ({
+    ...item,
+    isEdit: false,
+    children: item.children?addEditProperty(item.children):[] // 这里要判断原数据有没有子级如果没有判断会报错
+  }))
+
+
+  useEffect(() => {  
+    console.log("useEffect----openStatus")
+    if(openStatus>0){
+      loadingTree();
+      let initParam={
+        type:"type",
+        query:{DEVICE_STATUS:openStatus}
+      }
+      if(isAutoInitialized){
+        props.handleClick(initParam,openStatus,bindIndex);
+      } 
+    }else{
+      let initParam={
+        type:'list',
+        query:{}
+      }   
+      if(isAutoInitialized){
+        props.handleClick(initParam,openStatus,bindIndex);
+      }     
+      
+    } 
+    
+  }, [openStatus,bindIndex]);
+  useEffect(() => {  
+    console.log("useEffect----openStatus")
+    if(openStatus>0){
+      loadingTree();
+      let initParam={
+        type:"type",
+        query:{DEVICE_STATUS:openStatus}
+      }
+      if(isAutoInitialized){
+        props.handleClick(initParam,openStatus,bindIndex);
+      } 
+    }else{
+      let initParam={
+        type:'list',
+        query:{}
+      }   
+      if(isAutoInitialized){
+        props.handleClick(initParam,openStatus,bindIndex);
+      }     
+      
+    } 
+    
+  }, []);
+  const changeEditProperty = (arr,id,isEdit) => arr.map(item => {
+    item.isEdit= item.id===id?isEdit:false;
+    (
+    {
+    ...item,    
+    children: item.children?changeEditProperty(item.children,id,isEdit):[] // 这里要判断原数据有没有子级如果没有判断会报错
+  })})
+
   const operateCategory = (event) => {
-    console.log(event,selectNode); 
+    // debugger
+    console.log("operateCategory",selectNode); 
+    setCurValue("");
     if(event.key === 'delete'){
-      deleteAssetTree(selectNode["id"]).then(() =>{
-         loadingTree(openStatus);
-      })
+      // debugger;
+      Modal.confirm({
+        title: "确认要删除吗",
+        onOk: async () => {
+          console.log("delete)",openStatus);
+          deleteAssetTree(selectNode["id"]).then(() =>{
+            loadingTree();
+          })
+   
+        },
+        onCancel() { },
+      });
+
+
+
+      
+
+
     }else if(event.key === 'add'){
       let params = {
         name:"新建目录",
@@ -65,9 +148,12 @@ export default function Accordions(props: Props) {
         parent_id:selectNode["id"]
       }
       addAssetTree(params).then(() =>{
-        loadingTree(openStatus);
+        loadingTree();
      })
-    } 
+    }else{
+      changeEditProperty(treeData,selectNode["id"],true)
+    }
+    setShowMenu(false)
   }
   
   const renderMenu = () => {
@@ -78,6 +164,7 @@ export default function Accordions(props: Props) {
           style={{
             display: showMenu ? 'inherit' : 'none',
             position: 'fixed',
+            zIndex:1,
             left: pageX + 36,
             top: pageY + 2,
           }}
@@ -93,72 +180,122 @@ export default function Accordions(props: Props) {
     }
     return null;
   };
+
   const changeItem = (status,index) => {
       setBindIndex(index);
       setShowMenu(false);
-      setOpenStatus(status);     
-      if(status!=0){
-        loadingTree(status);
-        let node={
-          type:"type",
-          query:{DEVICE_STATUS:status}
-        }
-        props.handleClick(node,status,index);
-      }else{
-        let initParam={
-          type:'list',
-          query:{}
-        }
-        
-        props.handleClick(initParam,status,index);
-      }      
+      setOpenStatus(status);    
+           
   };
-  const initItem = (status,index) => {
-    setBindIndex(index);
-    setOpenStatus(status);     
-    if(status!=0){
-      loadingTree(status);
-    }     
-};
   const onSelect = (selectedKeys: React.Key[], info: any) => {
-    info.selectedNodes[0].status = openStatus;
     setShowMenu(false);
+    let query = info.selectedNodes[0];
+    console.log("debugger",query);
     props.handleClick(info.selectedNodes[0],openStatus,bindIndex);
   };
 
-  const loadingTree = (currentStatus) => {
-    getAssetsTree(currentStatus).then(res => {      
-      setTreeData(res.dat);
+  const loadingTree = () => {
+    
+    getAssetsTree(openStatus).then(({dat}) => {      
+    const treeData =  addEditProperty(dat);
+      setTreeData(treeData);
+      console.log("treeData",treeData);
     })
   };
 
-  useEffect(() => {
-     initItem(status,index);
-  }, [status,index]);
+
+  const saveNode = (id,parent_id) => {    
+    console.log("saveNode);",id,curValue);
+    let params = {
+      id:id,
+      name:curValue,
+      parent_id:parent_id,
+      status:openStatus,
+    }
+    updateAssetTree(params).then((res) => {
+      message.success("修改成功");
+      loadingTree()
+    })
+   
+  };
+
+  const onClose = (id) => {
+    changeEditProperty(treeData,id,false)
+    // setTreeData(treeData.slice());
+  };
 
 
   const titleRender = (node) => {
-      // console.log(node);
+      
       if(node.management_ip!=null && node.management_ip.length>0  &&  node.type=='asset'){
         node.name = node.management_ip;
       }else if( (node.name ==null || node.name=='') && node.type=='asset'){
         node.name = node.serial_number
-      }         
-      return (
-        <div style={{ position: 'relative', width: '100%' }}>
-            <span>{node.name}</span>
-             <span style={{ position: 'absolute', right: 5 }}>           
-          </span>
-        </div>
-      );
+      }   
+      // console.log("node", node.id,node.name);
+      if(node.isEdit){
+        return (
+          <div style={{ position: 'relative', width: '100%' }}>
+                <Input defaultValue={node.name}
+                maxLength={25}
+                style={{ width: '150px' }}
+                onChange={(e) => {
+                  setCurValue(e.target.value);
+                }}
+                onPressEnter={(e) => {                  
+                  saveNode(node.id,node.parent_id);
+                }}
+                onKeyDown={(e) => {
+                  if (e.code === 'Escape') {
+                   onClose(node);
+                  }
+                }}></Input>
+                <CloseOutlined
+                  style={{ marginLeft: 10 }}
+                  onClick={() => {
+                    onClose(node);
+                  }}
+                />
+                <CheckOutlined style={{ marginLeft: 10 }} onClick={() => saveNode(node.id,node.parent_id)} />
+                
+          </div>
+        );
+      }else{
+        return (
+          <div style={{ position: 'relative', width: '100%' }}>
+              <span>{node.name}</span>
+               <span style={{ position: 'absolute', right: 5 }}>           
+            </span>
+          </div>
+        );
+      }     
+      
   };
 
   const handleRightClick = ({ event, node }: any) => {
+    // debugger
     event.stopPropagation();
     setPageX(event.pageX);
     setPageY(event.pageY);
     setSelectNode(node);
-    setShowMenu(true);
+    if(node.type!=="asset"){
+      setShowMenu(true);
+    }else{
+      setShowMenu(false);
+    }    
+  };
+
+  const addTreeRootClick = ({ event}: any) => {
+    // debugger
+    // event.stopPropagation();
+    setPageX(event.pageX);
+    setPageY(event.pageY);
+    // setSelectNode(node);
+    // if(node.type!=="asset"){
+    //   setShowMenu(true);
+    // }else{
+    //   setShowMenu(false);
+    // }    
   };
 
   const addCatlog = (status,index)=>{
@@ -166,8 +303,9 @@ export default function Accordions(props: Props) {
       status:status,
       name:"新建目录"
     }
+    // debugger
     addAssetTree(params).then(()=>{
-      changeItem(status,index)
+      loadingTree()
     });
   }
 
@@ -188,16 +326,16 @@ export default function Accordions(props: Props) {
   return (
     <div className='bread-crumb-container'>
       {assetStatus.map(({ label,value }, i) => (
-        <div className='currentDiv' key={i} onClick={() => changeItem(value,i)} style={{ position: 'relative' }}>
+        <div className='currentDiv' key={i}  style={{ position: 'relative' }}>
           <div className='header'>
             <AppstoreAddOutlined/>
-            <span className='title' style={bindIndex != i ? style.noSelect : style.selected}>{label}</span>    
+            <span className='title'    onClick={() => changeItem(value,i)} style={bindIndex != i ? style.noSelect : style.selected}>{label}</span>    
                                  
             {i>0 && bindIndex == i && (  
               <>
               <span onClick={e=>{
                   addCatlog(value,i);
-               }} className='add_category' title='添加目录'> +</span>           
+               }} className='add_category' title='添加目录'  > +</span>           
               <DownOutlined className='jiantou0'/>
               </> 
             )}
